Refactor (WIP) Organized components
