+++ src/js/components/app-template.js
@@ -5,8 +5,8 @@
  */
 
 import React from 'react'
-import Header from './app-header'
-import Footer from './app-footer'
+import Header from './header/app-header'
+import Footer from './footer/app-footer'
 
 export default ( props ) => {
     return (
