+++ src/js/components/search/app-search.js
